Update docstrings and doctest

+++ py-polars/polars/series/datetime.py
@@ -274,7 +274,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -284,7 +284,7 @@ class DateTimeNameSpace:
         ... )
         >>> date.dt.quarter()
         shape: (4,)
-        Series: 'date' [u32]
+        Series: 'date' [i8]
         [
                 1
                 1
@@ -306,7 +306,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -316,7 +316,7 @@ class DateTimeNameSpace:
         ... )
         >>> date.dt.month()
         shape: (4,)
-        Series: 'date' [u32]
+        Series: 'date' [i8]
         [
                 1
                 2
@@ -338,7 +338,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -348,7 +348,7 @@ class DateTimeNameSpace:
         ... )
         >>> date.dt.week()
         shape: (4,)
-        Series: 'date' [u32]
+        Series: 'date' [i8]
         [
                 1
                 5
@@ -369,7 +369,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -377,7 +377,7 @@ class DateTimeNameSpace:
         >>> s = pl.date_range(date(2001, 1, 1), date(2001, 1, 7), eager=True)
         >>> s.dt.weekday()
         shape: (7,)
-        Series: 'date' [u32]
+        Series: 'date' [i8]
         [
                 1
                 2
@@ -402,7 +402,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -412,7 +412,7 @@ class DateTimeNameSpace:
         ... )
         >>> s.dt.day()
         shape: (5,)
-        Series: 'date' [u32]
+        Series: 'date' [i8]
         [
                 1
                 3
@@ -435,7 +435,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int16`.
 
         Examples
         --------
@@ -445,7 +445,7 @@ class DateTimeNameSpace:
         ... )
         >>> s.dt.ordinal_day()
         shape: (3,)
-        Series: 'date' [u32]
+        Series: 'date' [i16]
         [
                 1
                 32
@@ -558,7 +558,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -577,7 +577,7 @@ class DateTimeNameSpace:
         ]
         >>> date.dt.hour()
         shape: (4,)
-        Series: 'datetime' [u32]
+        Series: 'datetime' [i8]
         [
                 0
                 1
@@ -598,7 +598,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int8`.
 
         Examples
         --------
@@ -616,7 +616,7 @@ class DateTimeNameSpace:
         ]
         >>> date.dt.minute()
         shape: (3,)
-        Series: 'datetime' [u32]
+        Series: 'datetime' [i8]
         [
                 0
                 2
@@ -643,7 +643,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32` or :class:`Float64`.
+            Series of data type :class:`Int8` or :class:`Float64`.
 
         Examples
         --------
@@ -658,7 +658,7 @@ class DateTimeNameSpace:
         ... )
         >>> s.dt.second()
         shape: (3,)
-        Series: 'datetime' [u32]
+        Series: 'datetime' [i8]
         [
                 0
                 3
@@ -684,7 +684,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int32`.
 
         Examples
         --------
@@ -694,7 +694,7 @@ class DateTimeNameSpace:
         >>> s = pl.datetime_range(start, stop, interval="500ms", eager=True)
         >>> s.dt.millisecond()
         shape: (9,)
-        Series: 'datetime' [u32]
+        Series: 'datetime' [i32]
         [
                 0
                 500
@@ -718,7 +718,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int32`.
 
         Examples
         --------
@@ -742,7 +742,7 @@ class DateTimeNameSpace:
         ]
         >>> date.dt.microsecond()
         shape: (9,)
-        Series: 'datetime' [u32]
+        Series: 'datetime' [i32]
         [
                 0
                 500000
@@ -766,7 +766,7 @@ class DateTimeNameSpace:
         Returns
         -------
         Series
-            Series of data type :class:`UInt32`.
+            Series of data type :class:`Int32`.
 
         Examples
         --------
@@ -790,7 +790,7 @@ class DateTimeNameSpace:
         ]
         >>> date.dt.nanosecond()
         shape: (9,)
-        Series: 'datetime' [u32]
+        Series: 'datetime' [i32]
         [
                 0
                 500000000
